feat: Add mute log feature

+++ events/guild/messageDelete.js
@@ -4,7 +4,9 @@ import { deleteMessageHandler } from "#handlers"
 export default {
   name: Events.MessageDelete, 
   async execute(client, message) {
-    try {
+    if (message.author.bot) return
+	try {
+		
       await deleteMessageHandler(message);
     } catch (error) {
       console.error(`❌ Message Delete: `, error);
